migrate create dao related types into relevant components
