+++ apps/web/src/modules/create-dao/components/AllocationForm/ContributionAllocation.tsx
@@ -4,10 +4,10 @@ import React, { useState } from 'react'
 import AnimatedModal from 'src/components/Modal/AnimatedModal'
 import { PUBLIC_BUILDER_ADDRESS, PUBLIC_NOUNS_ADDRESS } from 'src/constants/addresses'
 import { useEnsData } from 'src/hooks/useEnsData'
-import { TokenAllocation } from 'src/typings'
 import { getEnsAddress } from 'src/utils/ens'
 
 import { useFormStore } from '../../stores'
+import { TokenAllocation } from '../AllocationForm'
 import { Contribution } from './Contribution'
 import {
   ContributionAllocationForm,
