+++ apps/web/src/modules/create-dao/components/AllocationForm/ContributionAllocationForm.tsx
@@ -5,9 +5,9 @@ import Date from 'src/components/Fields/Date'
 import SmartInput from 'src/components/Fields/SmartInput'
 import { PUBLIC_BUILDER_ADDRESS, PUBLIC_NOUNS_ADDRESS } from 'src/constants/addresses'
 import { useEnsData } from 'src/hooks/useEnsData'
-import { TokenAllocation } from 'src/typings'
 import { yearsAhead } from 'src/utils/helpers'
 
+import { TokenAllocation } from '../AllocationForm'
 import { validationSchemaContributions } from './AllocationForm.schema'
 import { DaoCopyAddress } from './DaoCopyAddress'
 import { Toggle } from './Toggle'
