+++ apps/web/src/modules/create-dao/stores/useFormStore.ts
@@ -2,14 +2,17 @@ import { create } from 'zustand'
 import { createJSONStorage, persist } from 'zustand/middleware'
 
 import { PUBLIC_BUILDER_ADDRESS, PUBLIC_NOUNS_ADDRESS } from 'src/constants/addresses'
+import { DaoContractAddresses } from 'src/modules/dao'
+import { yearsAhead } from 'src/utils/helpers'
+
 import {
   ArtworkFormValues,
   AuctionSettingsFormValues,
   GeneralFormValues,
-} from 'src/modules/create-dao'
-import { DaoContractAddresses } from 'src/modules/dao'
-import { IPFSUpload, OrderedLayersProps, TokenAllocation } from 'src/typings'
-import { yearsAhead } from 'src/utils/helpers'
+  IPFSUpload,
+  TokenAllocation,
+} from '../components'
+import { OrderedTraits } from '../components/Artwork/LayerBox'
 
 export interface FormStoreState {
   activeSection: number
@@ -32,8 +35,8 @@ export interface FormStoreState {
   setIpfsUpload: (ipfsUpload: IPFSUpload[]) => void
   deployedDao: DaoContractAddresses
   setDeployedDao: (deployedDao: DaoContractAddresses) => void
-  orderedLayers: OrderedLayersProps[]
-  setOrderedLayers: (orderedLayers: OrderedLayersProps[]) => void
+  orderedLayers: OrderedTraits
+  setOrderedLayers: (orderedLayers: OrderedTraits) => void
   isUploadingToIPFS: boolean
   setIsUploadingToIPFS: (bool: boolean) => void
   resetForm: () => void
@@ -114,7 +117,7 @@ export const useFormStore = create(
       setVetoPower: (vetoPower: boolean) => set({ vetoPower }),
       setSetUpArtwork: (artwork: ArtworkFormValues) => set({ setUpArtwork: artwork }),
       setIpfsUpload: (ipfsUpload: IPFSUpload[]) => set({ ipfsUpload }),
-      setOrderedLayers: (orderedLayers: OrderedLayersProps[]) => {
+      setOrderedLayers: (orderedLayers: OrderedTraits) => {
         set({
           orderedLayers,
         })
